Refactor: Updated Folder structure to follow best practices and screaming architecture
