+++ app/(tabs)/_layout.tsx
@@ -2,10 +2,10 @@ import { Tabs } from 'expo-router';
 import React from 'react';
 import { Platform } from 'react-native';
 
-import { HapticTab } from '@/components/HapticTab';
-import { IconSymbol } from '@/components/ui/IconSymbol';
-import TabBarBackground from '@/components/ui/TabBarBackground';
-import { Colors } from '@/constants/Colors';
+import { HapticTab } from '@/shared/components/HapticTab';
+import { IconSymbol } from '@/shared/components/ui/IconSymbol';
+import TabBarBackground from '@/shared/components/ui/TabBarBackground';
+import { Colors } from '@/shared/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
 export default function TabLayout() {
