+++ shared/components/HapticTab.tsx
