+++ shared/components/HelloWave.tsx
@@ -8,7 +8,7 @@ import Animated, {
   withSequence,
 } from 'react-native-reanimated';
 
-import { ThemedText } from '@/components/ThemedText';
+import { ThemedText } from '@/shared/components/ThemedText';
 
 export function HelloWave() {
   const rotationAnimation = useSharedValue(0);
