+++ shared/components/MessageBubble.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { View, StyleSheet } from 'react-native';
-import { ThemedText } from './ThemedText';
-import { Message } from '@/hooks/useChats';
-import { useColorScheme } from '@/hooks/useColorScheme';
+import { ThemedText } from '@/shared/components/ThemedText';
+import { Message } from '@/shared/hooks/db/useChatsDb';
+import { useColorScheme } from '@/shared/hooks/useColorScheme';
 
 interface MessageBubbleProps {
   message: Message;
@@ -25,7 +25,7 @@ export function MessageBubble({ message, isCurrentUser }: MessageBubbleProps) {
     ]}>
       <View style={[
         styles.bubble,
-        isCurrentUser 
+        isCurrentUser
           ? [styles.selfBubble, { backgroundColor: isDark ? '#235A4A' : '#DCF8C6' }]
           : [styles.otherBubble, { backgroundColor: isDark ? '#2A2C33' : '#FFFFFF' }]
       ]}>
