+++ shared/components/ui/TabBarBackground.ios.tsx
