+++ shared/components/ui/TabBarBackground.tsx
