+++ shared/database/schema/chats.ts
